% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{add_model_variables}
\alias{add_model_variables}
\title{Adds Additional Model Variables to Play-by-Play Data}
\usage{
add_model_variables(pbp_df)
}
\arguments{
\item{pbp_df}{Play-by-play data frame.}
}
\value{
Play-by-play data frame with the following columns added:
\itemize{
\item{"Shotgun_Ind"} - Indicator whether or not the play was in shotgun.
\item{"No_Huddle_Ind"} - Indicator whether or not the play was no huddle.
\item{"Home_Ind"} - Indicator whether or not the possession team was home.
\item{"airEPA_Result"} - airEPA for complete passes and EPA for incomplete.
\item{"airWPA_Result"} - airWPA for complete passes and WPA for incomplete.
\item{"yacEPA_Result"} - yacEPA for complete passes and EPA for incomplete.
\item{"yacWPA_Result"} - yacWPA for complete passes and WPA for incomplete.
\item{"Team_Side_Gap"} - Combine the team, side, and run gap for O-line proxy
}
}
\description{
Adds Additional Model Variables to Play-by-Play Data
}
\examples{
# Add model variables:
pbp_df <- pbp_df \%>\% add_model_variables()
}
