% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement_level_functions.R
\name{add_position_tables}
\alias{add_position_tables}
\title{Creates Player Tables for Each Position}
\usage{
add_position_tables(model_data_list)
}
\arguments{
\item{model_data_list}{List of two data frames (1) passing
play-by-play (pass_model_df) and (2) rushing play-by-play
(rush_model_df).}
}
\value{
List of dataframes with the given model_data_list,
as well as the tables for each position containing the
player names/IDs along with their respective number of
attempts for both passing/receiving and rushing/sacks.
}
\description{
Creates Player Tables for Each Position
}
\examples{
# Add the player tables to the model_data_list:
model_data_list <- model_data_list \%>\% add_position_tables()
}
