% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{add_positions}
\alias{add_positions}
\title{Adds Player Positions to Play-by-Play Data}
\usage{
add_positions(pbp_df, years)
}
\arguments{
\item{pbp_df}{Play-by-play data frame.}

\item{years}{Single number or vector of years to get team rosters for.}
}
\value{
Play-by-play data frame with the following columns added:
\itemize{
\item{"Passer_ID_Name"} - Name and ID of the passer
\item{"Receiver_ID_Name"} - Name and ID of the receiver
\item{"Rusher_ID_Name"} - Name and ID of the rusher
\item{"Passer_Position"} - Position of the passer
\item{"Receiver_Position"} - Position of the receiver
\item{"Rusher_Position"} - Position of the rusher
}
}
\description{
Adds Player Positions to Play-by-Play Data
}
\examples{
# Add positions for data from 2009 to 2016:
pbp_df <- pbp_df \%>\% add_positions(c(2009:2016))
}
