% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement_level_functions.R
\name{add_replacement_level_sim}
\alias{add_replacement_level_sim}
\title{Add Replacement Level Information to Simulation Data}
\usage{
add_replacement_level_sim(sim_data_list, replacement_tables)
}
\arguments{
\item{sim_data_list}{List of data frames in the simulation:
\itemize{
\item{"pass_model_df"} - Passing play-by-play data
\item{"rush_model_df"} - Rushing play-by-play data
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}

\item{replacement_tables}{List of position tables
that designate which players are replacement level:
\itemize{
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}
}
\value{
The input sim_data_list but with replacement-level
players' names in the play-by-play data all set to be
Replacement_POSITION and a new column for each of the
positional tables indicating which players are replacement
level.
}
\description{
Add Replacement Level Information to Simulation Data
}
\examples{
# Join positional replacement level information from
# the 2017 position tables to the simulated data:
sim_data_list <- sim_data_list \%>\%
  add_replacement_level_sim(replacement_tables_17)
}
