% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_stats_functions.R
\name{calc_passing_splits}
\alias{calc_passing_splits}
\title{Calculate Passing Statistics for Given Splits}
\usage{
calc_passing_splits(pass_pbp, splits = "Passer_ID_Name")
}
\arguments{
\item{pass_pbp}{Data frame of passing plays.}

\item{splits}{String vector of variables to split on, default
is "Passer_ID_Name" (splits are levels to calculate the
statistics for).}
}
\value{
Data frame of passing statistics evaluated at the
given level of splits.
}
\description{
Calculate Passing Statistics for Given Splits
}
\examples{
# Calculate passing statistics :
model_data_list <- calculate_passing_splits(model_data_list$pass_model_df,
                                            "Passer_ID_Name")
}
