% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_stats_functions.R
\name{calc_receiving_splits}
\alias{calc_receiving_splits}
\title{Calculate Receiving Statistics for Given Splits}
\usage{
calc_receiving_splits(pass_pbp, splits = "Receiver_ID_Name")
}
\arguments{
\item{pass_pbp}{Data frame of passing plays.}

\item{splits}{String vector of variables to split on, default
is "Receiver_ID_Name" (splits are levels to calculate the
statistics for).}
}
\value{
Data frame of receiving statistics evaluated at the
given level of splits.
}
\description{
Calculate Receiving Statistics for Given Splits
}
\examples{
# Calculate receiving statistics :
model_data_list <- calculate_receiving_splits(model_data_list$pass_model_df,
                                              "Receiver_ID_Name")
}
