% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/war_functions.R
\name{calculate_above_replacement}
\alias{calculate_above_replacement}
\title{Calculate Above Replacement Level}
\usage{
calculate_above_replacement(model_data_list)
}
\arguments{
\item{model_data_list}{List that must have the following:
\itemize{
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}
}
\value{
The input position tables with iPAR columns and
for air, yac, and rush along with their upper/lower bounds.
}
\description{
Calculate Above Replacement Level
}
\examples{
# Calculate the player's value above replacement level:
model_data_list <- model_data_list \%>\%
  calculate_above_replacement(model_data_list)
}
