% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/war_functions.R
\name{calculate_points_per_win}
\alias{calculate_points_per_win}
\title{Calculate the Points per Win for Given Season(s)}
\usage{
calculate_points_per_win(years)
}
\arguments{
\item{years}{Single number or vector of years to use in calculating
the number of points per win.}
}
\value{
Value indicating the points per win.
}
\description{
Calculate the Points per Win for Given Season(s)
}
\examples{
# Calculate the points per win in 2016:
points_per_win <- calculate_points_per_win(2016)
}
