% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_stats_functions.R
\name{calculate_rushing_splits}
\alias{calculate_rushing_splits}
\title{Calculate Rushing Statistics for Given Splits}
\usage{
calculate_rushing_splits(rush_pbp, splits = "Rusher_ID_Name")
}
\arguments{
\item{rush_pbp}{Data frame of rushing plays.}

\item{splits}{String vector of variables to split on, default
is "Rusher_ID_Name" (splits are levels to calculate the
statistics for).}
}
\value{
Data frame of rushing statistics evaluated at the
given level of splits.
}
\description{
Calculate Rushing Statistics for Given Splits
}
\examples{
# Calculate receiving statistics :
model_data_list <- calculate_rushing_splits(model_data_list$pass_model_df,
                                              "Rusher_ID_Name")
}
