% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{combine_simulations}
\alias{combine_simulations}
\title{Combine the Simulation Results}
\usage{
combine_simulations(sim_list)
}
\arguments{
\item{sim_list}{List of simulated results
where each member of the list has the following tables:
\itemize{
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}
}
\value{
List of the position tables but now with all
the simulation results rbind together.
}
\description{
Combine the Simulation Results
}
\examples{
sim_player_tables <- combine_simulations(sim_list)
}
