% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/war_functions.R
\name{convert_points_to_wins}
\alias{convert_points_to_wins}
\title{Convert Points to Wins}
\usage{
convert_points_to_wins(model_data_list, points_per_win)
}
\arguments{
\item{model_data_list}{List that must have the following:
\itemize{
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}

\item{points_per_win}{Value indicating the number of points
per win to use (will usually be the result of a function).}
}
\value{
The input position tables with WAR columns for
air, yac, rush, and totals along with upper/lower bounds.
}
\description{
Convert Points to Wins
}
\examples{
# Calculate the WAR values:
model_data_list <- model_data_list \%>\%
  convert_points_to_wins(model_data_list, points_per_win)
}
