% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/war_functions.R
\name{convert_prob_to_wins}
\alias{convert_prob_to_wins}
\title{Convert Probabilities to Wins}
\usage{
convert_prob_to_wins(model_data_list)
}
\arguments{
\item{model_data_list}{List that must have the following:
\itemize{
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}
}
\value{
The input position tables with WAR columns for
air, yac, rush, and totals along with upper/lower bounds.
}
\description{
Convert Probabilities to Wins
}
\examples{
# Calculate the WAR values:
model_data_list <- model_data_list \%>\%
  convert_prob_to_wins(model_data_list)
}
