% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement_level_functions.R
\name{create_percentage_replacement_fn}
\alias{create_percentage_replacement_fn}
\title{Creates Function to Find Percentage Based Replacement Level QBs}
\usage{
create_percentage_replacement_fn(attempt_type, attempt_percent)
}
\arguments{
\item{attempt_type}{String indicating which type of attempts to use from the table
to use for the percent cutoff, can be the following options: (1) Perc_Pass_Attempts,
(2) Perc_Rush_Attempts, (3) Perc_Sacks, or (4) Perc_Total_Plays.}

\item{attempt_percent}{Percentage of the given attempt_type for QBs on their team
that serves as a cutoff for replacement level.}
}
\value{
Function that only takes in the model_data_list and returns the
replacement level QBs.
}
\description{
Creates Function to Find Percentage Based Replacement Level QBs
}
\examples{
# First create the function to find the replacement QBs based on "Perc_Total_Plays"
# using a cutoff of 10\%:
find_replacement_QBs <- create_percentage_replacement_fn("Perc_Total_Plays", .1)
# Use this function to find the replacement level QBs:
replacement_QBs <- find_replacement_QBs(model_data_list)
}
