% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement_level_functions.R
\name{create_pos_team_table}
\alias{create_pos_team_table}
\title{Create non-QB table stats with team information}
\usage{
create_pos_team_table(model_data_list, position)
}
\arguments{
\item{model_data_list}{List of data frames:
\itemize{
\item{"pass_model_df"} - Passing play-by-play data
\item{"rush_model_df"} - Rushing play-by-play data
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}

\item{position}{String indicating which position to use, can be either
(1) RB/FB, (2) WR, or (3) TE.}
}
\value{
A different form of the position tables that are grouped by the player and team
level that also includes the player's number and percentage of different attempts with
respect to their team.
}
\description{
Create non-QB table stats with team information
}
\examples{
# Create the RB team table:
RB_team_table <- create_pos_team_table(model_data_list, "RB")
}
