% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement_level_functions.R
\name{create_QB_team_table}
\alias{create_QB_team_table}
\title{Create QB table stats with team information}
\usage{
create_QB_team_table(model_data_list)
}
\arguments{
\item{model_data_list}{List of data frames:
\itemize{
\item{"pass_model_df"} - Passing play-by-play data
\item{"rush_model_df"} - Rushing play-by-play data
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}
}
\value{
A different form of the QB table that is grouped by the player and team
level that also includes the player's number and percentage of different attempts with
respect to their team.
}
\description{
Create QB table stats with team information
}
\examples{
# Create the QB team table:
QB_team_table <- create_QB_team_table(model_data_list)
}
