% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_war_functions.R
\name{create_WAR_interval_plot}
\alias{create_WAR_interval_plot}
\title{Create WAR Interval Plots}
\usage{
create_WAR_interval_plot(war_tables, positions = c("QB", "RB", "FB", "WR",
  "TE"), war_type = "total")
}
\arguments{
\item{war_tables}{List that must have the following with the
various WAR calculations:
\itemize{
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}

\item{positions}{Vector of positions to include in the interval
plots (default is all positions QB, RB/FB, WR, TE).}

\item{war_type}{String indicating which type of war: (1) air,
(2) yac, (3) rush, (4) total, or (5) all - which plots the intervals
for the three different ones overlaid on each other}
}
\value{
Interval plot for the for input positions.
}
\description{
Create WAR Interval Plots
}
\examples{
# Generate the interval plots for QBs total WAR:
create_WAR_interval_plot(war_tables, positions = "QB", "total")
}
