% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{estimate_player_value_added}
\alias{estimate_player_value_added}
\title{Calculate the 3 Types iPA and iPAA Values for Each Player}
\usage{
estimate_player_value_added(model_data_list, model_formula_list, weight_type)
}
\arguments{
\item{model_data_list}{List of data frames:
\itemize{
\item{"pass_model_df"} - Passing play-by-play data
\item{"rush_model_df"} - Rushing play-by-play data
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}

\item{model_formula_list}{List of the following formulas:
\itemize{
\item{"air_formula"} - Formula for airEPA/airWPA model.
\item{"yac_forumla"} - Formula for yacEPA/yacWPA model.
\item{"qb_rush_formula"} - Formula for QB rushing/sacks model.
\item{"main_rush_formula"} - Formula for non-QB rushing model.
}}

\item{weight_type}{String indicating which type of observation
weighting to use, can either be: (1) "none" or (2) "scorediff".}
}
\value{
The input model_data_list but with each position table
modified so that the following columns are included:
\itemize{
\item{"air_iPA"} - Individual player points/probability added in the air.
\item{"yac_iPA"} - Individual player points/probability added after the catch.
\item{"rush_iPA"} - Individual player points/probability added from rushing.
\item{"air_iPAA"} - Individual player points/probability above average in the air.
\item{"yac_iPAA"} - Individual player points/probability above average after the catch.
\item{"rush_iPAA"} - Individual player points/probability above average from rushing.
}
as well as the resulting models:
\itemize{
\item{"air_model"} - Model fit for air component
\item{"yac_model"} - Model fit for yac component
\item{"qb_rush_model"} - Model fit for QB rushing/sacks
\item{"main_rush_model"} - Model for for all non-QB rushing attempts
}
}
\description{
Calculate the 3 Types iPA and iPAA Values for Each Player
}
\examples{
# Estimate the player effects and get the iPA and iPAA values:
model_data_list <- model_data_list \%>\%
  estimate_player_value_added(model_formula_list, "none")
}
