% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement_level_functions.R
\name{find_positional_replacement_level}
\alias{find_positional_replacement_level}
\title{Find Replacement Level for Each Position}
\usage{
find_positional_replacement_level(model_data_list, replacement_fns)
}
\arguments{
\item{model_data_list}{List of data frames:
\itemize{
\item{"pass_model_df"} - Passing play-by-play data
\item{"rush_model_df"} - Rushing play-by-play data
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}

\item{replacement_fns}{List of functions:
\itemize{
\item{"find_replacement_QB"} - Function to find replacement
level QBs
\item{"find_replacement_RB_rush"} - Function to find replacement
level RBs for rushing plays
\item{"find_replacement_WR_TE_rush"} - Function to find replacement
level TEs and WRs for rushing plays
\item{"find_replacement_WR_rec"} - Function to find replacement
level WRs for receiving plays
\item{"find_replacement_RB_rec"} - Function to find replacement
level RBs for receiving plays
\item{"find_replacement_TE_rec"} - Function to find replacement
level TEs for receiving plays
}}
}
\value{
The input model_data_list but with replacement-level
players' names in the play-by-play data all set to be
Replacement_POSITION and a new column for each of the
positional tables indicating which players are replacement
level.
}
\description{
Find Replacement Level for Each Position
}
\examples{
# Find positional replacement level:
model_data_list <- model_data_list \%>\%
  find_positional_replacement_level(replacement_functions)
}
