% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{get_season_summary}
\alias{get_season_summary}
\title{Creates Summary of Team Games}
\usage{
get_season_summary(years)
}
\arguments{
\item{years}{Single number or vector of years to get team rosters for.}
}
\value{
Data frame with a row every team's record and score differential
in each of the given years.
}
\description{
Creates Summary of Team Games
}
\examples{
# Summary of team performances in 2009:
season_summary_09 <- get_season_summary(2009)
}
