% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_stats_functions.R
\name{join_position_statistics}
\alias{join_position_statistics}
\title{Join Various Statistics to the Position Tables}
\usage{
join_position_statistics(model_data_list)
}
\arguments{
\item{model_data_list}{List that must have the following:
\itemize{
\item{"pass_model_df"} - Passing play-by-play data
\item{"rush_model_df"} - Rushing play-by-play data
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
}}
}
\value{
The input position tables with all the different
types of statistics included.
}
\description{
Join Various Statistics to the Position Tables
}
\examples{
# Calculate the various types of statistics for each position:
model_data_list <- model_data_list \%>\%
  join_position_statistics(model_data_list)
}
