% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{measure_meta_independence}
\alias{measure_meta_independence}
\title{Generate the Independence R-Squared Matrix for Given Metrics}
\usage{
measure_meta_independence(stat_seasons, attempt_col, min_attempts, stat_cols)
}
\arguments{
\item{stat_seasons}{Data frame of statistics with the
provided player_col, season_col, and stat_cols.}

\item{attempt_col}{String indicating the name of the column
to filter the data on.}

\item{min_attempts}{Number of minimum of the attempt_col
that a player must have to be included.}

\item{stat_cols}{String vector of the statistics for which
the independence matrix will be generated for.}
}
\value{
The independence R-squared matrix.
}
\description{
Generate the Independence R-Squared Matrix for Given Metrics
}
\examples{
qb_independence <- measure_meta_independence(qb_seasons,
"Total_Plays", 100, c("total_WAR", "Comp_Perc", "Success_Rate"))
}
