% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{measure_meta_reliability}
\alias{measure_meta_reliability}
\title{Evaluate the Discrimination and Stability for Given Metrics}
\usage{
measure_meta_reliability(boot_seasons, stat_seasons, player_col, season_col,
  attempt_col, min_attempts, stat_cols)
}
\arguments{
\item{boot_seasons}{Data frame of bootstrapped statistics
with the provided player_col, season_col, and stat_cols.}

\item{stat_seasons}{Data frame of statistics with the
provided player_col, season_col, and stat_cols.}

\item{player_col}{String indicating the name of the column
in the data sets standing for the player name.}

\item{season_col}{String indicating the name of the column
in the data sets standing for the season.}

\item{attempt_col}{String indicating the name of the column
to filter the data on.}

\item{min_attempts}{Number of minimum of the attempt_col
that a player must have to be included.}

\item{stat_cols}{String vector of the statistics for which
the discrimination and stability will be calculated for.}
}
\value{
Data frame with three columns (1) Statistic - which
is the name of one of the stats, (2) Discrimination - the
average discrimination of the Statistic over all provided
seasons, (3) Stability - the stability for the Statistic
over all the provided seasons.
}
\description{
Evaluate the Discrimination and Stability for Given Metrics
}
\examples{
# Calculate the discrimination and stability for QB stats:
qb_reliability <-
measure_meta_reliability(qb_boot_seasons,qb_seasons,
"Player_ID_Name", "Season", "Total_Plays", 100,
c("total_WAR", "Comp_Perc", "Success_Rate"))
}
