% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{prepare_model_data}
\alias{prepare_model_data}
\title{Filter the Play-by-Play Data to Data for Modeling}
\usage{
prepare_model_data(pbp_df)
}
\arguments{
\item{pbp_df}{Play-by-play data frame.}
}
\value{
List of two data frames: (1) passing play-by-play
data, and (2) rushing play-by-play data that has been
filtered to only include plays used for modeling.
}
\description{
Filter the Play-by-Play Data to Data for Modeling
}
\examples{
# Create list of model data:
model_data_list <- pbp_df \%>\% prepare_model_data()
}
