% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{resample_season}
\alias{resample_season}
\title{Resample Season}
\usage{
resample_season(pbp_data, drive_level = 0)
}
\arguments{
\item{drive_level}{Indicator whether or not to resample
entire drives in the season (default is 0).}

\item{pbp_df}{Play-by-play data frame.}
}
\description{
Resample Season
}
\examples{
# Resample the 2017 season:
resampled_pbp_17 <- resample_season(get_pbp_data(2017))
}
