% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{simulate_player_value_added}
\alias{simulate_player_value_added}
\title{Simulate to Estimate Player Effects Variance and Intervals}
\usage{
simulate_player_value_added(model_data_list, n_sim = 1000)
}
\arguments{
\item{model_data_list}{List of data frames:
\itemize{
\item{"pass_model_df"} - Passing play-by-play data
\item{"rush_model_df"} - Rushing play-by-play data
\item{"QB_table"} - Table of QBs
\item{"RB_table"} - Table of RBs
\item{"WR_table"} - Table of WRs
\item{"TE_table"} - Table of TEs
\item{"air_model"} - Model fit for air component
\item{"yac_model"} - Model fit for yac component
\item{"qb_rush_model"} - Model fit for QB rushing/sacks
\item{"main_rush_model"} - Model for for all non-QB rushing attempts
}}

\item{n_sim}{Number of simulations (default is 1000)}
}
\value{
The input model_data_list but with each position table
modified so that upper/lower estimates for each iPA and iPAA
value are estimated from +/- two standard deviations from
the simulations.
}
\description{
Simulate to Estimate Player Effects Variance and Intervals
}
\examples{
# Simulate the player effects and get intervals for the iPA and iPAA values:
model_data_list <- model_data_list \%>\%
  simulate_player_value_added(model_data_list, n_sim = 1000)
}
