% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{estimate_rushing_value_added}
\alias{estimate_rushing_value_added}
\title{Calculate Player Effects for Rushing Plays}
\usage{
estimate_rushing_value_added(rush_pbp_df, qb_rush_formula, main_rush_formula)
}
\arguments{
\item{rush_pbp_df}{Play-by-play data set of passing plays.}

\item{qb_rush_formula}{Formula for QB rushing/sacks.}

\item{main_rush_formula}{Formula for non-QB rushing.}
}
\value{
List with the following:
\itemize{
\item{"qb_rush_model"} - Model fit for air component
\item{"main_rush_model"} - Model fit for yac component
\item{"QB_table"} - Table of QB estimated random effects
\item{"Rush_table"} - Table of rusher estimated random effects
}
}
\description{
Calculate Player Effects for Rushing Plays
}
\examples{
rushing_effects_list <- estimate_rushing_value_added(rush_model_df, qb_rush_formula,
                                                     main_rush_formula, "scorediff")
}
