% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement_level_functions.R
\name{find_percentage_replacement_QB}
\alias{find_percentage_replacement_QB}
\title{Find Replacement-Level QBs based on the Percentage of Plays Cutoff}
\usage{
find_percentage_replacement_QB(qb_team_table, attempt_type, attempt_percent)
}
\arguments{
\item{qb_team_table}{Data frame with QB attempts in relation to their team}

\item{attempt_type}{String indicating which type of attempts to use from the table
to use for the percent cutoff.}

\item{attempt_percent}{Percentage of the given attempt_type for QBs on their team
that serves as a cutoff for replacement level.}
}
\value{
Vector of Player_ID_Name values indicating the replacement level
QBs given the inputs.
replacement level for the position given the league replacement_depth and attempt_type.
}
\description{
Find Replacement-Level QBs based on the Percentage of Plays Cutoff
}
\examples{
# Find the replacement level QBs based on 10\% cutoff for Total_Plays:
replacement_QBs <- find_percentage_replacement_QB(model_data_list, "Total_Plays", .1)
}
