% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{get_pbp_data}
\alias{get_pbp_data}
\title{Retrieves nflscrapR Play-by-Play Data}
\usage{
get_pbp_data(years, live = FALSE)
}
\arguments{
\item{years}{Single number or vector of years to get play-by-data for.}

\item{live}{Indicator for whether or not to scrape the data or access
saved files on github. Defaults to FALSE.}
}
\value{
Data frame of the play-by-play for the given year(s).
}
\description{
Retrieves nflscrapR Play-by-Play Data
}
\examples{
# Get data from 2009 to 2016:
pbp_df <- get_pbp_data(c(2009:2016))
}
